Adapters.
Connect with more than just `http`; now initial support for `express` and `hapi`. It made sense to include first-class functionality for doing this since it will probably be a common pattern and will make migration easier. Documentation has been updated where appropriate.

Special updates to the test scripts had to be made for `hapi` since it doesn't support `node@0.12`. When this case is detected, the `hapi` tests are not run.