Remove support for `webpack-udev-server`.
The `assets` middleware is now primarily built for just the simple serving of the webpack build output folder (with some fast checks on file existence). The dev-server middleware is now part of `webpack-udev-server` itself, and the two exported functions in the `assets` middleware can be used to construct any middleware from which assets may be served.