Prevent `upgrade` default handler with other listeners present.
There isn't really a "catch-all" like `res.finished` for the `upgrade` event. So if we're the only listener then we know we can close the connection. Otherwise we just pray whomever else has attached to the event knows what they're doing. Such is the case when using things like `socket.io`.