Add `fetch` testing utility.
The `fetch` utility can be used to test your `midori` apps easily without having to spin up a dedicated server for every test. It works mostly like the usual what-wg `fetch` except it takes a midori app or app creator as its first argument and the promise result is closer to a node `ServerResponse` than anything else. Usage highlights can be found in `send.spec.js` and `server.spec.js` for now, as those two relied most heavily on the actual body content of the response.