Invoke `listening` if server is already listening.
If you connect an app that has a `listening` handler you would expect this to be called if the server is listening. I guess this deviates from the normal spec a little but this feels largely like intended behaviour. Especially if you wish to do things like setup or teardown.