Allow mocking of selectors during tests.
 * Add `createMockRequest`, `createMockResponse` test functions.
 * Add `runSelector` and `getSelectorImplementation` test functions.

Instead of running through the entire request lifecycle, `runSelector` allows you to just test the value and execution of a single selector, while mocking out all the selectors it depends on. Examples and usage can be found in the README. `createMockRequest` and `createMockResponse` have been extracted to allow the mocking of the `request` "selector".